mtd: Introduce SPI NAND framework
This includes following changes :-

1. mtd: nand: Add JEDEC manufacturer ID for Gigadevice
This commit adds Gigadevice to the list of manufacturer ID and name strings.

2. mtd: nand: Allow to set a per-device ECC layout
Until now, the ECC was handled either by Linux (i.e. software ECC) or by the
NAND controller (i.e. hardware ECC). In each of these, the ECC layout is
defined by the controller of the NAND device, not by the NAND device itself.

However, devices with on-die ECC support have their own ECC layout so we need
a way to specify a per-device ECC layout.

This commit adds a new field to the nand_flash_dev structure, to allow devices
to specify its own ECC layout.

3. mtd: nand: Check length of ID before reading bits per cell
The table-based NAND identification currently reads the number
of bits per cell from the 3rd byte of the extended ID. This is done
for the so-called 'full ID' devices; i.e. devices that have a known
length ID.

However, if the ID length is shorter than three, there's no 3rd byte,
and so it's wrong to read the bits per cell from there. Fix this by
adding a check for the ID length.

4. mtd: Introduce SPI NAND framework
Add a new framework, to support SPI NAND devices. The framework registers
a NAND chip and handles the generic SPI NAND protocol, calling device-specific
hooks for each SPI NAND command.

The following is the stack design, from userspace to hardware. This commit
adds the "SPI NAND core" layer.

    Userspace
  ------------------
    MTD
  ------------------
    NAND core
  ------------------
    SPI NAND core
  ------------------
    SPI NAND device
  ------------------
    SPI core
  ------------------
    SPI master
  ------------------
    Hardware

(backported from https://patchwork.ozlabs.org/patch/416854/)

5. mtd: spi-nand: Support common SPI NAND devices
This commit uses the recently introduced SPI NAND framework to support
Micron MT29F and Gigadevice GD5F serial NAND devices. These two families
of devices are fairly similar and so they are supported with only minimal
differences.

The current support includes:

  * Page read and page program operations (using on-die ECC)
  * Page out-of-band read
  * Erase
  * Reset
  * Device status retrieval
  * Device ID retrieval

(am from https://patchwork.ozlabs.org/patch/416855/)

6. mtd: spi-nand: Return number of corrected bitflips when reading
This commit fixes a bug in the current SPI NAND implementation, which
misses the update of the 'ecc_stats.corrected' field in the MTD main structure.

This field holds the number of corrected ECC bitflips, as reported by
the device's internal ECC controller.

Without this patch, the bitflips aren't reported to upper layers such
as UBI. UBI uses the number of corrected bitflips to decide if an
eraseblock needs scrubbing.

7. mtd: spi-nand: Set page third byte address on block erase
The SPI NAND block erase command expects the page-aligned address of
the block's first page. However, this is currently broken as the
implementation only sets the lower two bytes. Fix it.

8. mtd: spi-nand: Rework SPI NAND command transfer
Using a separate struct spi_transfer for the command and the address
results in random I/O errors. This commit reworks the way SPI messages
are prepared for each command, putting the command and the address
in a single struct spi_transfer.

While here, add some sanity checks on the command.

9. mtd: spi-nand: Remove "wait until ready" in the write path
The "wait until ready" call reads the STATUS (0xC0) register, and waits
until the OIP (Operation-In-Progress) flag is cleared.

Currently, it is used after page program and block erase commands are sent,
to check they are completed and the device is idle.

The call is also used between the cache programand the page program
commands. However, the GD5F specification does not mention it as part
of the write sequence. Instead, the write sequence is specified as:

1) Program Load (cache program)
2) Write Enable
3) Program Execute (page program)
4) Read Status register and wait until ready

Instensive testing shows it's actually problematic, causing pages to not get
properly written. Remove it.

The transfer mode, obtained from the
DTS file is checked, and the transfer command is determined based
on the obtained value. This way, to switch between single, dual
and quad modes only the DTS file has to be modified by altering
the RX/TX bus width value.

Signed-off-by: Abhimanyu Vishwakarma <Abhimanyu.Vishwakarma@imgtec.com>
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>