clk: pistachio: Implement PLL rate adjustment
This commit implements small rate changes to the fractional PLL.
This is done using the PLL frac parameter. The .set_rate function
first finds the parameters associated to the closest nominal rate.

Then the new rate is set, using parameters from the table entry,
except for the frac parameter, which is calculated from the rate
using the fractional PLL rate formula.

Using .round_rate, the driver guarantees that only rates near
a table nominal rate is applied. To this extent, add two parameters
fout_min and fout_max, which allows to define the allowed rate
adjustment.

Also includes following changes :-

1. clk: pistachio: Extend DIV_F to pass clk_flags as well
As preparation work to support MIPS PLL rate change propagation, this
commit extends the DIV_F macro to pass clk_flags in addition to div_flags.

2. clk: pistachio: Add a MUX_F macro to pass clk_flags
As preparation work to support MIPS PLL rate change propagation, this
commit adds a MUX_F macro to pass clk_flags.

3. clk: pistachio: Propagate rate changes in the MIPS PLL clock sub-tree
This commit passes CLK_SET_RATE_PARENT to the "mips_div", "mips_internal_div",
and "mips_pll_mux" clocks. This flag is needed for the "mips" clock to
propagate rate changes up to the "mips_pll" root clock.

4. clk: pistachio: Add a rate table for the MIPS PLL
This commit adds a rate parameter table, which makes it possible for
the MIPS PLL to support rate change.

5. clocksource: Add Pistachio clocksource-only driver
The Pistachio SoC provides four general purpose timers, and allow
to implement a clocksource driver.

This driver can be used as a replacement for the MIPS GIC and MIPS R4K
clocksources and sched clocks, which are clocked from the CPU clock.

Given the general purpose timers are clocked from an independent clock,
this new clocksource driver will be useful to introduce CPUFreq support
for Pistachio machines.

6. clk-pistachio : adding wifi_pll rate table

7. pistachio: pll: Fix 32bit integer overflows
This commit fixes 32bit integer overflows throughout the pll driver
(i.e. wherever the result of integer multiplication may exceed the
range of u32).

One of the functions affected by this problem is .recalc_rate. It
returns incorrect rate for some pll settings (not for all though)
which in turn results in the incorrect rate setup of pll's child
clocks.

8. pistachio: pll: Fix pll rate calculation in integer mode
Bug: .recalc_rate callback for the fractional pll doesn't take operating
mode into account when calculating pll rate. This bug results in
the incorrect pll rates when pll is operating in integer mode.

This commit fixes this bug by checking the operating mode before
calculating pll rate.

BUG=none
TEST=Pistachio BuB

9. pistachio: pll: Set operating mode in .set_rate
This commit sets operating mode of fractional pll based on
the value of the fractional divider. Previous version assumed that
the pll will always be configured in fractional mode which may not be
the case. This may result in the wrong output frequency.

10. pistachio: pll: Fix vco calculation in .set_rate (fractional)

Fix the bug in vco frequency calculation in .set_rate for fractional
pll. Vco was calculated based on the current operating mode which
makes no sense because .set_rate is setting operating mode. Instead,
vco should be calculated using pll settings that are about to be set.
This bug was introduced in commit d259b8083fb6b51ad9c431ed73d1c745af4eaf08,
but it doesn't affect clock tree setup.

Change-Id: I429c15e6c76c4f8c4e34a5cdc6e08f8858a705f2
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Signed-off-by: Zdenko Pulitika <zdenko.pulitika@imgtec.com>
Signed-off-by: Sridhar Nuvusetty <Sridhar.Nuvusetty@imgtec.com>

pistachio: cpu-freq: update freq table.

Updating mips pll rate table to enter
lower frequencies.

Current freq table with on-demand governor
sets low freq to 416Mhz, with this update
the governor will try to put it to 208Mhz.

Change-Id: Ie90456372a43679a4e70a0896fa893125698e39a
Signed-off-by: Govindraj Raja <govindraj.raja@imgtec.com>