drivers/misc: add U-Boot bootcount driver
This driver implements the Linux kernel half of the boot count feature -
the boot counter can only be reset after it is clear that the
application has been started and is running correctly, which usually
can only be determined by the application code itself. Thus the reset
of the boot counter must be done by application code, which thus needs
an appropriate driver.

This driver provides read/write access to the U-Boot bootcounter
through sysfs file.

Signed-off-by: Avinash Tahakik <Avinash.Tahakik@imgtec.com>
Signed-off-by: Shraddha Chaudhari <Shraddha.Chaudhari@imgtec.com>