UBIFS: use flush_kernel_dcache_page instead of flush_dcache_page
When the kernel needs to modify a user page it has obtained with
kmap, it should call this function  after all modifications are
complete, but before kunmapping it, to bring L2 caches and underlying
memory up to date.
This is the behavior we see for UBIFS read/writepage and therefore
flush_kernel_dcache_page should be used instead of flush_dcache_page.
By comparison, flush_dcache_page is used at any time the kernel
writes to a page in the page cache or reads from a page in the page cache
with potential shared/writable userspace mappings.

Change-Id: Ie017e8718f87c326f7f7f58f803a38dd5bf0b70c
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>