spi: spidev: Increase default value of bufsize to 65535
It was not possible to change the parameter bufsiz while loading
the spidev driver because OpenWRT does not support adding parameters
to AutoProbe.
The spi-img-spfi driver supports transfer length up to 65535 bytes
but spidev default limit is 4096 bytes. This commit changes the limit
of spidev to match the limit of spi-img-spfi driver.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>