spi: img-spfi: Implement quad-mode for transmit
This requires changes to the SPFI driver consisting of setting
the TX_RX bit which lets the state machine know that the command
which is going to be sent to the flash is a TX (set)/RX(cleared)
command. This helps with the scenario in which we have both a
command transfer (command, address, dummy) which will be pending(TX),
and a TX data transfer, which was problematic before. This bit of
the control register is flash specific and has no significance in
SPI traditional sense.

Change-Id: Ic21752a6ae8a304f52fd46d9ffebd90c93b388c8
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>